move MyHomePage to a separate file and update routing
